Move executor to pyquil_utils.
