+++ mitiq/mitiq_pyquil/tests/test_zne_mitiq_pyquil.py
@@ -26,10 +26,9 @@ from mitiq.zne import (
     mitigate_executor,
     zne_decorator,
 )
-from mitiq.mitiq_pyquil.executor import (
-    generate_qcs_executor, ground_state_expectation
-)
 from mitiq.mitiq_pyquil.pyquil_utils import (
+    generate_qcs_executor,
+    ground_state_expectation,
     random_identity_circuit,
 )
